+++ TypeScript/src/gilded-rose.ts
@@ -1,7 +1,8 @@
 import { Item } from './item/Item'
-import { decreaseQualityToMin0, increaseQualityToMax50, expire } from './item/quality'
+import { decreaseQualityToMin0, increaseQualityToMax50 } from './item/quality'
 import { backstagePass, agedBrie, normal } from './item/itemTypes'
 import { decreaseSellIn, expired } from './item/sellIn'
+import { handleBackstagePass } from './handleBackstagePass'
 
 export class GildedRose {
   constructor(private items: Item[] = []) {}
@@ -26,16 +27,4 @@ function handleNormalItem(item: Item) {
 function handleBrie(item: Item) {
   decreaseSellIn(item)
   increaseQualityToMax50(item, expired(item) ? 2 : 1)
-}
-
-function handleBackstagePass(item: Item) {
-  increaseBackstageQuality(item)
-  decreaseSellIn(item)
-  if (expired(item)) expire(item)
-}
-
-function increaseBackstageQuality(item: Item) {
-  if (item.sellIn < 6) increaseQualityToMax50(item, 3)
-  else if (item.sellIn < 11) increaseQualityToMax50(item, 2)
-  else increaseQualityToMax50(item)
 }
